Improves stability of options page.
A service worker now correctly checks for, and loads default options
when the extension is first installed or updated.
The options page now asks for confirmation from the user when trying to
leave with unsaved changes.
There is now a button to restore the extension's default settings, with
propper confirmation.
Adds implementation of accessible announcement system into the options
page and replaces alerts() with this system.